Implemented Delete with Dialog

+++ src/app/components/list-user/list-user.component.ts
@@ -1,9 +1,12 @@
 
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { Dialog } from '@angular/cdk/dialog';
+import { Component, Inject, OnInit, ViewChild } from '@angular/core';
+import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
 import { User } from 'src/app/model/user';
+import { DialogService } from 'src/app/services/dialog.service';
 import { UserService } from 'src/app/services/user.service';
 
 @Component({
@@ -25,7 +28,7 @@ export class ListUserComponent implements OnInit {
     this.dataSource.paginator = this.paginator;
   }
 
-  constructor(private userService: UserService, private router: Router) { }
+  constructor(private userService: UserService, private router: Router, private dialogService: DialogService) { }
 
   ngOnInit(): void {
     this.getData();
@@ -38,13 +41,28 @@ export class ListUserComponent implements OnInit {
   }
 
   deleteUser(userId: number) {
-    this.userService.deleteById(userId).subscribe(
-      res => { this.dataSource.data = res }
-    )
+    // this.userService.deleteById(userId).subscribe(
+    //   res => { this.dataSource.data = res }
+    // )
+    this.userService.findById(userId).subscribe({
+      next: user => {
+        this.currentUser = user;
+      }
+    });
+
+    this.dialogService.openConfirmDialog(`Vuoi cancellare ${this.currentUser?.nome} ${this.currentUser?.cognome}?`)
+      .afterClosed().subscribe(res => {
+        if (res) {
+          this.userService.deleteById(userId).subscribe(
+            result => { this.dataSource.data = result }
+          );
+        }
+      });
+
   }
 
   editUser(userId: number) {
-    throw new Error('Method not implemented.');
+    this.router.navigate(['edit/', userId])
   }
 
   showUser(userId: number) {
